feat(v3): project structure migration — v2 → v3 layout
- Extract core/finalizer.py from brain.py (hallucination check, LLM finalization)
- Extract memory/nodes.py (node/relationship schema, entity extraction)
- Extract memory/context_builder.py (graph → LLM context string)
- Move TUI to interface/tui/ with panels/ sub-package + styles.tcss
- Move Telegram bot to interface/telegram_bot.py
- Backward-compat shims in old locations (app.py, integrations/)
- Remove empty ui/ directory (replaced by interface/tui/)
- Update all imports throughout

Closes #81
