+++ src/bantz/interface/__init__.py
@@ -1,0 +1,1 @@
+"""Bantz — Interface layer (TUI + Telegram)."""
