+++ src/bantz/interface/tui/__init__.py
@@ -1,0 +1,1 @@
+"""Bantz — Textual TUI package."""
