+++ src/bantz/interface/tui/panels/__init__.py
@@ -1,0 +1,5 @@
+"""Bantz — TUI Panel widgets."""
+from bantz.interface.tui.panels.system import SystemStatus
+from bantz.interface.tui.panels.chat import ChatLog, ThinkingLabel
+
+__all__ = ["SystemStatus", "ChatLog", "ThinkingLabel"]
