+++ src/bantz/memory/__init__.py
@@ -1,1 +1,7 @@
-"""Bantz — Graph Memory (Neo4j)."""
+"""Bantz — Graph Memory (Neo4j).
+
+Sub-modules:
+    nodes           — node/relationship schema + entity extraction
+    context_builder — graph query → LLM-ready context string
+    graph           — Neo4j connection layer + GraphMemory singleton
+"""
