+++ src/bantz/memory/graph.py
@@ -25,23 +25,10 @@ import re
 from datetime import datetime
 from typing import Optional
 
+from bantz.memory.nodes import NODE_LABELS, REL_TYPES, extract_entities
+from bantz.memory.context_builder import build_context
+
 log = logging.getLogger("bantz.graph")
-
-# ── Node / Relationship constants ─────────────────────────────────────────
-
-NODE_LABELS = ("Person", "Topic", "Decision", "Task", "Event", "Location", "Document")
-
-REL_TYPES = {
-    "KNOWS":        ("Person", "Person"),
-    "ASSIGNED_TO":  ("Task", "Person"),
-    "RELATED_TO":   (None, None),          # flexible
-    "DECIDED_IN":   ("Decision", "Event"),
-    "WORKS_ON":     ("Person", "Topic"),
-    "LOCATED_AT":   (None, "Location"),
-    "REFERENCES":   (None, "Document"),
-    "COMMITTED_TO": ("Person", "Task"),
-    "FOLLOWS_UP":   ("Task", "Decision"),
-}
 
 
 class GraphMemory:
@@ -146,185 +133,8 @@ class GraphMemory:
         tool_used: str | None,
         tool_data: dict | None,
     ) -> list[dict]:
-        """
-        Rule-based entity extraction from conversation.
-        Returns list of {"label": ..., "props": {...}, "rels": [...]}
-        """
-        entities: list[dict] = []
-        combined = f"{user_msg} {assistant_msg}".lower()
-        tool_data = tool_data or {}
-
-        # ── People mentioned ──
-        # Look for name patterns: "Ali said", "meeting with Ahmet", "email from John"
-        people_patterns = [
-            r"(?:from|with|to|by|about|ask|tell|call|email|meet)\s+([A-Z][a-z]{2,15}(?:\s+[A-Z][a-z]{2,15})?)",
-        ]
-        found_people: set[str] = set()
-        for pat in people_patterns:
-            for m in re.finditer(pat, f"{user_msg} {assistant_msg}"):
-                name = m.group(1).strip()
-                # Filter out common non-name words
-                skip = {"Monday", "Tuesday", "Wednesday", "Thursday", "Friday",
-                        "Saturday", "Sunday", "January", "February", "March",
-                        "April", "May", "June", "July", "August", "September",
-                        "October", "November", "December", "Today", "Tomorrow",
-                        "Error", "Done", "Event", "Calendar", "Gmail", "News",
-                        "Desktop", "Downloads", "Documents", "Home", "Bantz",
-                        "Linux", "Python", "English", "Turkish"}
-                if name not in skip and len(name) > 2:
-                    found_people.add(name)
-
-        for name in found_people:
-            entities.append({
-                "label": "Person",
-                "key": "name",
-                "props": {"name": name, "last_seen": datetime.now().isoformat()},
-                "rels": [],
-            })
-
-        # ── Topics from tool usage ──
-        tool_topic_map = {
-            "calendar": "calendar",
-            "gmail": "email",
-            "news": "news",
-            "weather": "weather",
-            "classroom": "university",
-            "web_search": "research",
-            "document": "documents",
-            "shell": "system",
-        }
-        if tool_used and tool_used in tool_topic_map:
-            topic_name = tool_topic_map[tool_used]
-            entities.append({
-                "label": "Topic",
-                "key": "name",
-                "props": {"name": topic_name, "last_accessed": datetime.now().isoformat()},
-                "rels": [],
-            })
-
-        # ── Calendar events → Event nodes ──
-        if tool_used == "calendar" and tool_data:
-            events = tool_data.get("events", [])
-            if isinstance(events, list):
-                for ev in events[:5]:
-                    title = ev.get("summary") or ev.get("title", "")
-                    if title:
-                        start = ev.get("start", "")
-                        entities.append({
-                            "label": "Event",
-                            "key": "title",
-                            "props": {"title": title, "date": start,
-                                      "updated_at": datetime.now().isoformat()},
-                            "rels": [],
-                        })
-            # Created event from message
-            if "added" in assistant_msg.lower() or "created" in assistant_msg.lower():
-                title_m = re.search(r"[\"'](.+?)[\"']|Event:\s*(.+?)(?:\n|$)", assistant_msg)
-                if title_m:
-                    t = (title_m.group(1) or title_m.group(2) or "").strip()
-                    if t:
-                        entities.append({
-                            "label": "Event",
-                            "key": "title",
-                            "props": {"title": t, "updated_at": datetime.now().isoformat()},
-                            "rels": [],
-                        })
-
-        # ── Gmail → potential person nodes ──
-        if tool_used == "gmail" and tool_data:
-            messages = tool_data.get("messages", [])
-            for msg in messages[:5]:
-                sender = msg.get("from", "")
-                if sender:
-                    # Extract name from "Name <email>" format
-                    name_match = re.match(r"([^<]+)", sender)
-                    if name_match:
-                        name = name_match.group(1).strip().strip('"')
-                        if name and len(name) > 2 and "@" not in name:
-                            entities.append({
-                                "label": "Person",
-                                "key": "name",
-                                "props": {"name": name, "email": sender.strip(),
-                                          "last_seen": datetime.now().isoformat()},
-                                "rels": [],
-                            })
-
-        # ── Decisions — "let's use X", "we'll go with X", "decided to X" ──
-        decision_patterns = [
-            r"(?:let'?s|we(?:'ll)?\s+(?:go\s+with|use|pick|choose|decided?\s+to))\s+(.+?)(?:\.|$|!)",
-            r"(?:i'?ll|going to|plan to)\s+(.+?)(?:\.|$|!)",
-        ]
-        for pat in decision_patterns:
-            m = re.search(pat, combined)
-            if m:
-                what = m.group(1).strip()[:100]
-                if len(what) > 5:
-                    entities.append({
-                        "label": "Decision",
-                        "key": "what",
-                        "props": {"what": what,
-                                  "date": datetime.now().isoformat(),
-                                  "context": user_msg[:200]},
-                        "rels": [],
-                    })
-                break  # one decision per exchange
-
-        # ── Tasks — "remind me to", "i need to", "add task" ──
-        task_patterns = [
-            r"(?:remind\s+me\s+to|i\s+need\s+to|todo|add\s+task|task:)\s+(.+?)(?:\.|$|!)",
-            r"(?:don'?t\s+forget\s+to)\s+(.+?)(?:\.|$|!)",
-        ]
-        for pat in task_patterns:
-            m = re.search(pat, combined)
-            if m:
-                desc = m.group(1).strip()[:150]
-                if len(desc) > 3:
-                    entities.append({
-                        "label": "Task",
-                        "key": "description",
-                        "props": {"description": desc, "status": "open",
-                                  "priority": "medium",
-                                  "created_at": datetime.now().isoformat()},
-                        "rels": [],
-                    })
-                break
-
-        # ── Documents ──
-        if tool_used == "document":
-            path = tool_data.get("path", "")
-            if path:
-                entities.append({
-                    "label": "Document",
-                    "key": "path",
-                    "props": {"path": path,
-                              "accessed_at": datetime.now().isoformat()},
-                    "rels": [],
-                })
-
-        # ── Locations ──
-        loc_patterns = [
-            r"(?:in|at|from|going to|travel to)\s+([A-Z][a-z]+(?:\s+[A-Z][a-z]+)?)"
-            r"(?:\s|,|\.|\?|!|$)",
-        ]
-        skip_locs = {"Monday", "Tuesday", "Wednesday", "Thursday", "Friday",
-                     "Saturday", "Sunday", "January", "February", "March",
-                     "April", "May", "June", "July", "August", "September",
-                     "October", "November", "December", "Today", "Tomorrow",
-                     "Error", "Done", "English", "Turkish", "Bantz", "Ollama",
-                     "Gmail", "Google", "Linux", "Python", "Neo"}
-        for pat in loc_patterns:
-            for m in re.finditer(pat, f"{user_msg} {assistant_msg}"):
-                loc = m.group(1).strip()
-                if loc not in skip_locs and len(loc) > 2:
-                    entities.append({
-                        "label": "Location",
-                        "key": "name",
-                        "props": {"name": loc,
-                                  "last_mentioned": datetime.now().isoformat()},
-                        "rels": [],
-                    })
-
-        return entities
+        """Delegate to bantz.memory.nodes module."""
+        return extract_entities(user_msg, assistant_msg, tool_used, tool_data)
 
     async def _upsert_entities(self, entities: list[dict], now: str) -> None:
         """MERGE entities into Neo4j — create if new, update if existing."""
@@ -352,68 +162,10 @@ class GraphMemory:
     # ── Reading — build context for LLM ────────────────────────────────────
 
     async def context_for(self, user_msg: str) -> str:
-        """
-        Query the graph for entities relevant to the user's message
-        and return a compact multi-line string the LLM can use as context.
-        """
+        """Delegate to bantz.memory.context_builder module."""
         if not self._enabled:
             return ""
-
-        try:
-            parts: list[str] = []
-
-            # 1. Recent people
-            people = await self._query(
-                "MATCH (p:Person) RETURN p.name AS name, p.last_seen AS seen "
-                "ORDER BY p.last_seen DESC LIMIT 8"
-            )
-            if people:
-                names = [r["name"] for r in people]
-                parts.append(f"Known people: {', '.join(names)}")
-
-            # 2. Open tasks
-            tasks = await self._query(
-                "MATCH (t:Task {status: 'open'}) "
-                "RETURN t.description AS desc, t.priority AS prio "
-                "ORDER BY t.created_at DESC LIMIT 5"
-            )
-            if tasks:
-                lines = [f"  - [{t['prio']}] {t['desc']}" for t in tasks]
-                parts.append("Open tasks:\n" + "\n".join(lines))
-
-            # 3. Recent decisions
-            decisions = await self._query(
-                "MATCH (d:Decision) RETURN d.what AS what, d.date AS date "
-                "ORDER BY d.date DESC LIMIT 3"
-            )
-            if decisions:
-                lines = [f"  - {d['what']}" for d in decisions]
-                parts.append("Recent decisions:\n" + "\n".join(lines))
-
-            # 4. Recent events
-            events = await self._query(
-                "MATCH (e:Event) RETURN e.title AS title, e.date AS date "
-                "ORDER BY e.updated_at DESC LIMIT 5"
-            )
-            if events:
-                lines = [f"  - {e['title']} ({e['date'] or '?'})" for e in events]
-                parts.append("Recent events:\n" + "\n".join(lines))
-
-            # 5. Keyword-matching — find nodes whose names/descriptions match user query
-            keywords = self._extract_keywords(user_msg)
-            if keywords:
-                relevant = await self._keyword_search(keywords)
-                if relevant:
-                    parts.append("Related context:\n" + "\n".join(f"  - {r}" for r in relevant))
-
-            if not parts:
-                return ""
-
-            return "=== Graph Memory ===\n" + "\n".join(parts) + "\n=== End Graph ==="
-
-        except Exception as exc:
-            log.debug("Graph context error: %s", exc)
-            return ""
+        return await build_context(user_msg, self._query)
 
     async def _query(self, cypher: str, **params) -> list[dict]:
         """Run a read query, return list of record dicts."""
@@ -421,39 +173,6 @@ class GraphMemory:
             result = await s.run(cypher, **params)
             records = await result.data()
             return records
-
-    def _extract_keywords(self, text: str) -> list[str]:
-        """Pull significant words from user input for graph search."""
-        stop = {"the", "a", "an", "is", "are", "was", "were", "do", "does",
-                "did", "will", "can", "could", "would", "should", "my", "me",
-                "i", "you", "we", "it", "he", "she", "they", "this", "that",
-                "what", "how", "when", "where", "who", "which", "to", "in",
-                "on", "at", "for", "of", "and", "or", "but", "not", "with",
-                "from", "by", "about", "any", "all", "some", "up", "out",
-                "just", "now", "today", "tomorrow", "yesterday", "please",
-                "check", "show", "tell", "get", "want", "need", "have", "has",
-                "hey", "hi", "hello", "thanks", "ok", "okay", "yeah", "yes", "no"}
-        words = re.findall(r"[a-zA-Z]{3,}", text.lower())
-        return [w for w in words if w not in stop][:5]
-
-    async def _keyword_search(self, keywords: list[str]) -> list[str]:
-        """Search graph nodes for keyword matches."""
-        results: list[str] = []
-        for kw in keywords:
-            pattern = f"(?i).*{re.escape(kw)}.*"
-            rows = await self._query(
-                "MATCH (n) WHERE any(k IN keys(n) WHERE "
-                "toString(n[k]) =~ $pattern) "
-                "RETURN labels(n)[0] AS label, "
-                "coalesce(n.name, n.title, n.description, n.what, n.path) AS val "
-                "LIMIT 3",
-                pattern=pattern,
-            )
-            for r in rows:
-                entry = f"[{r['label']}] {r['val']}"
-                if entry not in results:
-                    results.append(entry)
-        return results[:8]
 
     # ── Stats ──────────────────────────────────────────────────────────────
